+++ packages/hardhat-deploy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midl-xyz/hardhat-deploy
 
+## 2.0.0-next.48
+
+### Patch Changes
+
+- feat(satoshi-kit): expose account and network to verify and sign message
+  feat(executor): getEVMAddress is now possible to use w/o midl config
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@2.0.0-next.48
+
 ## 2.0.0-next.47
 
 ### Patch Changes